partner_ssn AFC-1641: module added for merge to Dev-12.0 partner_view_360 AFC-1641: removed anything found in partner_ssn/l10n_se, added dependency on l10n_se
