+++ partner_ssn/views/res_partner_view.xml
@@ -1,0 +1,16 @@
+<odoo>
+
+    <record id="view_partner_form" model="ir.ui.view">
+        <field name="name">res.partner.skill</field>
+        <field name="model">res.partner</field>
+        <field name="inherit_id" ref="base.view_partner_form"/>
+        <field name="arch" type="xml">
+            <field name="category_id" position="before">
+                <field name="social_sec_nr" />
+            </field>
+        </field>
+    </record>
+
+
+
+</odoo>
